feat(api): be able assign a list of cluster&namespce to a project (#137)
We used to assign a k8s namespace (for all clusters) to a project,
but this changes API so that we can assign a namespace of a specific
cluster to a project.

The new "assignments" field is added to CreateProject and Project, but
the backward compatibility is kept. The "kubernetes_namespace" field
is populated from the "assignments" field.